Move non-capistrano dependent stuff into lib/fiesta
