+++ lib/capistrano/fiesta.rb
@@ -1,4 +1,4 @@
-require "capistrano/fiesta/repo_url_parser"
-require "capistrano/fiesta/report"
+require "fiesta/repo_url_parser"
+require "fiesta/report"
 
 load File.expand_path("../tasks/fiesta.rake", __FILE__)
