--- lib/capistrano/fiesta/logger.rb
@@ -1,16 +1,0 @@
-module Capistrano
-  module Fiesta
-    class Logger
-      @logs = []
-
-      def self.warn(message)
-        @logs << "[FIESTA] #{message}"
-        nil
-      end
-
-      def self.logs
-        @logs
-      end
-    end
-  end
-end
