--- lib/capistrano/fiesta/timestamp_normalizer.rb
@@ -1,18 +1,0 @@
-require "attr_extras/explicit"
-
-module Capistrano
-  module Fiesta
-    class TimestampNormalizer
-      extend AttrExtras.mixin
-      pattr_initialize :timestamp
-
-      def run
-        if timestamp.is_a?(Time)
-          timestamp
-        else
-          Time.parse(timestamp.to_s + "Z")
-        end
-      end
-    end
-  end
-end
