+++ lib/fiesta/logger.rb
@@ -1,0 +1,14 @@
+module Fiesta
+  class Logger
+    @logs = []
+
+    def self.warn(message)
+      @logs << "[FIESTA] #{message}"
+      nil
+    end
+
+    def self.logs
+      @logs
+    end
+  end
+end
