+++ lib/fiesta/templates/draft.erb
