+++ test/auto_composed_story_test.rb
@@ -1,6 +1,6 @@
 require "test_helper"
 
-module Capistrano::Fiesta
+module Fiesta
   class AutoComposedStoryTest < Minitest::Test
     def test_only_returning_images_with_special_anchor
       pr = OpenStruct.new(body: "one pic http://github.com/avatar.jpg and another http://google.com/fish.png#")
