+++ test/test_helper.rb
@@ -2,8 +2,8 @@
 ENV["OCTOKIT_ACCESS_TOKEN"] = "token-value"
 
 # Require files to test
-require 'capistrano/fiesta/repo_url_parser'
-require 'capistrano/fiesta/report'
+require 'fiesta/repo_url_parser'
+require 'fiesta/report'
 
 # Additional tooling
 require 'minitest/autorun'
